# Conflicts:
#	client/src/main/java/nl/tudelft/oopp/demo/communication/BikeReservationCommunication.java
#	client/src/main/java/nl/tudelft/oopp/demo/communication/FoodServerCommunication.java
#	client/src/main/java/nl/tudelft/oopp/demo/entities/BikeReservation.java
#	client/src/main/java/nl/tudelft/oopp/demo/entities/Food.java
#	server/src/main/java/nl/tudelft/oopp/demo/controllers/BikeReservationController.java
#	server/src/main/java/nl/tudelft/oopp/demo/controllers/BuildingController.java
#	server/src/main/java/nl/tudelft/oopp/demo/entities/BikeReservation.java
#	server/src/main/java/nl/tudelft/oopp/demo/entities/Food.java
#	server/src/main/java/nl/tudelft/oopp/demo/entities/FoodReservations.java
#	server/src/main/java/nl/tudelft/oopp/demo/repositories/FoodRepository.java

+++ client/src/main/java/nl/tudelft/oopp/demo/communication/FoodServerCommunication.java
@@ -14,7 +14,7 @@ public class FoodServerCommunication {
      * @param name The name of the new food
      * @param price The price of the new food
      */
-    public static void createFood(String name, double price) {
+    public static void createFood(String name, int price) {
         String params = "name=" + name + "&price=" + price;
         sendPost("createFood", params);
     }
@@ -56,7 +56,7 @@ public class FoodServerCommunication {
      * @param name The new name of the food
      * @param price The new price of the food
      */
-    public static void updateFood(int id, String name, double price) {
+    public static void updateFood(int id, String name, int price) {
         String params = "id=" + id + "&name=" + name + "&price=" + price;
         sendPost("updateFood", params);
     }
